[MSE] Media segment is incorrectly dropped when using negative timestampOffset or when source buffer appendWindow is set. https://bugs.webkit.org/show_bug.cgi?id=222260
Patch by Jean-Yves Avenard <jya@apple.com> on 2021-02-24
Reviewed by Eric Carlson.

Source/WebCore:

Test: media/media-source/media-source-timestampoffset-trim.html

CoreMedia packs multiple audio frames together into a single CMSampleBuffer,
this allows for faster processing and easier insertion into the track buffer tree.
However, per mediasoure spec [1], a frame is to be dropped according to
its start time and duration. So if only the beginning of the MediaSample
was to be dropped, we would have incorrectly dropped the lot.
We now split the MediaSample if it is going to be dropped to ensure that
all usable content is inserted into the track buffer.
Audio splicing isn't done yet, but this gets us closer to it.

[1] https://w3c.github.io/media-source/#sourcebuffer-coded-frame-processing

* platform/graphics/SourceBufferPrivate.cpp:
(WebCore::SourceBufferPrivate::didReceiveSample):
* platform/graphics/SourceBufferPrivate.h:
* platform/graphics/avfoundation/objc/SourceBufferPrivateAVFObjC.mm:
(WebCore::SourceBufferPrivateAVFObjC::didParseInitializationData):
(WebCore::SourceBufferPrivateAVFObjC::didProvideMediaDataForTrackId):
* platform/graphics/gstreamer/mse/SourceBufferPrivateGStreamer.cpp:
(WebCore::SourceBufferPrivateGStreamer::didReceiveSample):
* platform/graphics/gstreamer/mse/SourceBufferPrivateGStreamer.h:

LayoutTests:

* media/media-source/media-source-timestampoffset-trim.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273461 268f45cc-cd09-0410-ab3c-d52691b4dbfc