Streamline ImageData size calculations and handle out-of-memory https://bugs.webkit.org/show_bug.cgi?id=222340
Reviewed by Geoff Garen.

* html/ImageData.cpp:
(WebCore::ImageData::dataSize): Added
(WebCore::ImageData::create): Use the dataSize function consistently to compute
the size of the byteArray needed. In the JavaScript-exposed create function, use
tryCreateUninitialized and throw an exception if it fails. Removed an unnecessary
assertion that simply repeats something covered by an if statement just above it.
(WebCore::ImageData::ImageSize): Removed the overload that only takes an IntSize
and allocates a Uint8ClampedArray, moving that into the create function instead.

* html/ImageData.h: Added dataSize, removed a constructor overload.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273505 268f45cc-cd09-0410-ab3c-d52691b4dbfc