Length blending should allow for a ValueRange parameter https://bugs.webkit.org/show_bug.cgi?id=222497
Reviewed by Dean Jackson.

We introduced a new NonNegativeLengthPropertyWrapper in r273001, the fix for bug 222034, with a special
blend() override implementation which would disallow negative values. However, this is something that
should ideally be provided lower down in the blending code in the blend() function on the Length type
itself. This will allow us to blend other non-negative length-related types, such as LengthBox, without
duplicating this code.

We also use this opportunity to not use the { 0, Fixed } value in all cases where we clip to 0, and
instead use the type from the blended values, provided one isn't 0.

* animation/CSSPropertyAnimation.cpp:
(WebCore::blendFunc):
(WebCore::NonNegativeLengthPropertyWrapper::NonNegativeLengthPropertyWrapper):
* platform/Length.cpp:
(WebCore::blend):
* platform/Length.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273603 268f45cc-cd09-0410-ab3c-d52691b4dbfc