[JSC] WebAssembly: make Wasm::Signature 32bit friendly https://bugs.webkit.org/show_bug.cgi?id=222543
Patch by Xan Lopez <xan@igalia.com> on 2021-03-01
Reviewed by Yusuke Suzuki.

The Wasm code uses the address of a Signature object as its
index. To make this work in 32bit just change Wasm::SignatureIndex
to be a uintptr_t instead of uint64_t. Also, remove some
unnecessary includes while we are at it.

* wasm/WasmModule.h:
* wasm/WasmSignature.h:
* wasm/js/JSWebAssemblyModule.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273662 268f45cc-cd09-0410-ab3c-d52691b4dbfc