EventDispatcher::wheelEvent is accessing m_scrollingTrees without locking m_scrollingTreesMutex since r271235 https://bugs.webkit.org/show_bug.cgi?id=222529
Reviewed by Carlos Garcia Campos.

r271235 added a code to access m_scrollingTrees, but it didn't
lock m_scrollingTreesMutex. The code can be removed by moving it
into the above lambda expression.

EventDispatcher::wheelEvent was using a local lambda expression.
However the return value wasn't used. And, MSVC can't compile the
lambda expression because it has a bug that 'this' can't be used
in lambda expressions. Converted the lambda expression to a
do-while statement.

* WebProcess/WebPage/EventDispatcher.cpp:
(WebKit::EventDispatcher::wheelEvent):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273669 268f45cc-cd09-0410-ab3c-d52691b4dbfc