Crash under KeyframeEffect::setTarget() https://bugs.webkit.org/show_bug.cgi?id=222591 <rdar://problem/74281295>
Reviewed by David Kilzer.

The Styleable returned by targetStyleable() holds a reference to the Element that at
this point is m_target (assuming it's non-null). However, once we set the new value
for m_target, if the only reference to the original target was held by this KeyframeEffect,
then that element will be deallocated and by the time we call didChangeTargetStyleable()
it will be gone.

To address, we create a RefPtr<Element> in the scope of KeyframeEffect::setTarget()
protecting the Styleable's element while didChangeTargetStyleable() is called.

* animation/KeyframeEffect.cpp:
(WebCore::KeyframeEffect::setTarget):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@273752 268f45cc-cd09-0410-ab3c-d52691b4dbfc